+++ lib/utils/network/http/http_screen_client.dart
@@ -4,6 +4,7 @@ import 'dart:io';
 
 // import 'package:chucker_flutter/chucker_flutter.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:http/http.dart' as http;
 import 'package:lokal/configs/environment.dart';
 import 'package:lokal/constants/strings.dart';
@@ -17,7 +18,7 @@ import '../../../constants/json_constants.dart';
 import '../../../pages/UikBottomNavigationBar.dart';
 import '../../../screens/Onboarding/OnboardingScreen.dart';
 import '../../storage/user_data_handler.dart';
-
+import 'package:internet_connection_checker/internet_connection_checker.dart';
 class HttpScreenClient {
   // static HttpClient getHttp() {
   //   return http.Client();
@@ -40,7 +41,7 @@ class HttpScreenClient {
                   child: const Text(LOG_IN),
                   onPressed: () {
                     UserDataHandler.clearUserToken();
-                    NavigationUtils.pop();
+                    Navigator.of(context).pop();
                     Navigator.pushAndRemoveUntil(
                       NavigationUtils.getCurrentContext()!,
                       MaterialPageRoute(
@@ -89,20 +90,25 @@ class HttpScreenClient {
   }
 
   static Future<ApiResponse> getApiResponse(String pageRoute, args) async {
-    var bodyParams = (args != null) ? args : <String, dynamic>{};
-    var header = NetworkUtils.getRequestHeaders();
     try {
+      bool hasConnection = await InternetConnectionChecker().hasConnection;
+      if (!hasConnection) {
+        displayNoInternetDialog(null);
+        throw Exception('No internet connection');
+      }
+
+      var bodyParams = args ?? <String, dynamic>{};
+      var header = NetworkUtils.getRequestHeaders();
       final response = await http.Client()
           .post(
-            Uri.parse(Environment().config.BASE_URL + pageRoute),
-            headers: header,
-            body: jsonEncode(bodyParams),
-          )
+        Uri.parse(Environment().config.BASE_URL + pageRoute),
+        headers: header,
+        body: jsonEncode(bodyParams),
+      )
           .timeout(Duration(seconds: NetworkUtils.REQUEST_TIMEOUT));
 
       if (response.statusCode == NetworkUtils.HTTP_SUCCESS) {
-        ApiResponse apiResponse =
-            ApiResponse.fromJson(jsonDecode(response.body));
+        ApiResponse apiResponse = ApiResponse.fromJson(jsonDecode(response.body));
         if (apiResponse.isSuccess!) {
           return apiResponse;
         } else {
@@ -112,26 +118,58 @@ class HttpScreenClient {
               case NetworkUtils.NETWORK_ERROR_USER_NOT_AUTHENTICATED:
                 {
                   displayUserUnAuthorisedDialog();
-                  throw Exception('Failed to load $pageRoute');
+                  throw Exception('User not authenticated');
                 }
-                break;
               default:
                 {
                   UiUtils.showToast(apiResponse.error![MESSAGE]);
-                  throw Exception('Failed to load $pageRoute');
+                  throw Exception('API response error');
                 }
             }
           } else {
-            throw Exception('Failed to load $pageRoute');
+            throw Exception('API response error');
           }
         }
       } else {
-        throw Exception('Failed to load $pageRoute');
+        throw Exception('API request failed');
       }
     } on TimeoutException catch (_) {
-      throw Exception('Timeout Error to load $pageRoute');
+      throw Exception('Timeout Error');
     } on SocketException catch (_) {
-      throw Exception('Socket Error to load $pageRoute');
-    }
+      throw Exception('Socket Error');
+    } catch (e, stackTrace) {
+      // Log the error for debugging and monitoring
+      debugPrint('API request failed: $e'+ stackTrace.toString());
+      throw e;
+     }
   }
+
+
+  static displayNoInternetDialog(Function? retryCallback) {
+    return showDialog(
+      barrierDismissible: false,
+      context: NavigationUtils.getCurrentContext()!,
+      builder: (BuildContext context) {
+        return WillPopScope(
+          onWillPop: () => Future.value(false),
+          child: AlertDialog(
+            title: const Text("No Internet Connecton !"),
+            actions: <Widget>[
+              MaterialButton(
+                color: Colors.amberAccent,
+                textColor: Colors.black,
+                child: const Text("Close App"),
+                onPressed: () {
+                  Navigator.of(context).pop(); // Close the dialog
+                  SystemNavigator.pop();
+                 // retryCallback(); // Call the retry callback
+                },
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+
 }