+++ lib/utils/network/retrofit/api_routes.dart
@@ -73,12 +73,11 @@ class ApiRoutes {
       "/service/getAllAgentsForUserService";
   static const String getServiceDetailsById = "/service/getDetailsById";
   static const String addPartnerAgent = "/partnerAgent/add";
-
   static const String notifyAllAgents = "/notifyAllAgents";
-
   static const String getAllUserAgentByPartnerId =
       "/user/getAllUserAgentByPartnerId";
-
   static const String createOrUpdateForAgents =
       "/userService/createOrUpdateForAgents";
+  static const String getAgentDetailsByPartnerIdAndServiceId =
+      "/partnerAgent/getAgentDetailsByPartnerIdAndServiceId";
 }